requirements.txt
google-generativeai==0.8.3
cloudscraper
PyMuPDF==1.24.0